added some fixes in repositories added some widgets

+++ lib/src/models/coordinate.dart
@@ -1,17 +1,17 @@
 class Coordinate {
-  int? x;
-  int? y;
+  double? x;
+  double? y;
 
   Coordinate({this.x, this.y});
 
-  factory Coordinate.fromList(List<int> l) {
+  factory Coordinate.fromList(List<double> l) {
     return Coordinate(
       x: l.first,
       y: l.last,
     );
   }
 
-  List<int> toList() {
+  List<double> toList() {
     return [x ?? 0, y ?? 0];
   }
 }
